Refactor odometry into nodelet
