+++ goldorak_base/launch/base.launch
@@ -1,20 +1,33 @@
 <launch>
+  <arg name="simulated_can" default="false"/>
+
   <rosparam file="$(find goldorak_bringup)/param/uavcan_nodes.yaml"/>
   <rosparam file="$(find goldorak_base)/param/base.yaml"/>
 
+  <!-- Nodelet manager -->
   <node pkg="nodelet" type="nodelet" name="base_nodelet_manager"
         args="manager"
         output="screen">
-    <param name="num_worker_threads" value="4"/>
+    <param name="num_worker_threads" value="8"/>
   </node>
 
-  <node pkg="nodelet" type="nodelet" name="uavcan_bridge_nodelet"
-        args="load goldorak_base/uavcan_bridge_nodelet base_nodelet_manager"
+  <!-- Uavcan bridge nodelet -->
+  <group unless="$(arg simulated_can)">
+    <node pkg="nodelet" type="nodelet" name="uavcan_bridge_nodelet"
+          args="load goldorak_base/uavcan_bridge_nodelet base_nodelet_manager"
+          output="screen">
+    </node>
+  </group>
+
+  <!-- Joint state publisher nodelet -->
+  <node pkg="nodelet" type="nodelet" name="joint_state_publisher_nodelet"
+        args="load goldorak_base/joint_state_publisher_nodelet base_nodelet_manager"
         output="screen">
   </node>
 
-  <node pkg="nodelet" type="nodelet" name="joint_state_publisher_nodelet"
-        args="load goldorak_base/joint_state_publisher_nodelet base_nodelet_manager"
+  <!-- Odometry nodelet -->
+  <node pkg="nodelet" type="nodelet" name="odometry_nodelet"
+        args="load goldorak_base/odometry_nodelet base_nodelet_manager"
         output="screen">
   </node>
