+++ goldorak_base/nodelets.xml
@@ -17,3 +17,13 @@
     </description>
   </class>
 </library>
+
+<library path="lib/libodometry_nodelet">
+  <class name="goldorak_base/odometry_nodelet"
+         type="goldorak_base::odometry_nodelet"
+         base_class_type="nodelet::Nodelet">
+    <description>
+      Goldorak's differential wheelbase odometry
+    </description>
+  </class>
+</library>
